requirements.txt
tensorflow==2.18.0
tensorflow-serving-api==2.18.0
python-multipart==0.0.9
pillow==10.3.0
matplotlib==3.8.4
numpy==1.26.4
tensorflow-addons==0.21.0
tensorflow-model-optimization==0.8.0
matplotlib==3.8.4
scikit-learn==1.3.2
pandas
seaborn
statsmodels
streamlit